+++ History.md
@@ -1,3 +1,16 @@
+0.4.1 / 2015-02-08
+==================
+
+* set connect timeout for polling
+
+0.4.0 / 2015-01-25
+==================
+
+* compatible with engine.io 1.5.1
+* fix default port detection when host is specified
+* add `Socket#id()` for retrieving `id`
+* make `Socket.priorWebsocketSuccess` private
+
 0.3.1 / 2014-11-04
 ==================